Some kind of refactoring, for better memory usage
Image files were opened for almost every request using the imaging
library. Because of this it was started to eat the memory, by reading
the file over again and again.

Now only opens the image only if need to convert/resize it.
Somehow it still not freeing the memory for the image as soon as
possible, but much-much better.